Fix clippy deny: cast_possible_truncation / cast_possible_wrap / cast_sign_loss
Stage 2 of the clippy deny-flip series. Eliminates all remaining
integer-cast lints by replacing `as` with `try_from` + saturating
fallback, explicit `cast_signed()` for u32 -> libc::pid_t /
u64 -> i64 conversions, and `From::from` for the few truly
infallible widenings that still used `as`.

Notable patterns:
- TUI geometry (`usize::len() as u16`) uses u16::try_from with
  `u16::MAX` saturation. This never truncates in practice because
  the inputs are bounded by terminal width (well under 65k), but
  the lint requires a deliberate narrowing construct and the
  saturation is the honest one.
- libc::killpg takes `pid_t = i32`; callers use
  `child.id().cast_signed()` and `pid.cast_signed()` so the pid
  -> pid_t reinterpret is explicit.
- Duration::as_secs() -> i64 in the metrics aggregator uses
  .cast_signed() - no clock we model reaches 2^63 seconds so the
  wrap is not observable.
- Added `usize_to_u16_sat` helper in ui/detail_pane.rs for the
  repeated UnicodeWidthStr::width pattern; the pattern was used 5
  times in that file alone.
- Mouse protocol byte encoding (src/event/mouse/selection.rs)
  previously compared `cx > 255 || cy > 255` then narrowed with
  `as u8`; now uses a `match (u8::try_from(cx), u8::try_from(cy))`
  pair so the narrowing and the bound check are the same
  operation.
- percentile index calculation in dashboard/kpis.rs was already
  rewritten in Stage 1 to pure integer arithmetic; no f64 path
  remains.

No functional changes; every conversion is semantically identical
to the `as` it replaced, just expressed through a safer API.

+++ src/ui/dashboard/kpis.rs
@@ -44,7 +44,7 @@ pub fn draw_dashboard_kpis(
         .get(&from_day)
         .copied()
         .unwrap_or(backlog_now);
-    let delta = backlog_now as i32 - backlog_then as i32;
+    let delta = i64::from(backlog_now) - i64::from(backlog_then);
     let delta_str = if delta >= 0 {
         format!("+{delta}")
     } else {
